requirements.txt
openai>=1.0.0
google-api-python-client
google-auth
google-auth-oauthlib
google-generativeai
anthropic
Flask
neo4j
python-dotenv
numpy
scikit-learn
gspread
oauth2client
requests
pandas
flask-cors
flask-jwt-extended
flask-socketio
